requirements.txt
# Core Django Dependencies
Django==5.0.4
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.3.0
asgiref>=3.8.1
sqlparse>=0.4.4
tzdata>=2023.3

# Environment Variables
python-dotenv==1.0.0

# Real-time WebSocket Support
channels==4.0.0
daphne==4.0.0

# Optional: Redis for production (uncomment if needed)
# channels-redis==4.1.0